requirements.txt
compressai==1.2.4
einops==0.8.0
kornia==0.7.0
lpips==0.1.4
numpy==1.23.1
omegaconf==2.3.0
open_clip_torch==2.22.0
openai_clip==1.0.1
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.4.0
pyiqa==0.1.7
pytorch_lightning==1.5.0
thop==0.1.1.post2209072238
timm==0.9.7
torchvision
tqdm==4.66.1
transformers==4.41.0
ultralytics_thop==2.0.0
# xformers==0.0.22
